requirements.txt
Django==2.0.4
djangorestframework==3.8.2
coreapi==2.3.3
openapi-codec==1.3.2
simplejson==3.9.0
django-filter

# Testing
tox==2.4.1
pylint==1.8.4
coverage==4.5.1
ipdb==0.10.1

# Docs
mkdocs==0.15.3

# Heroku
dj_database_url==0.5.0
psycopg2==2.7.4
gunicorn==19.8.0
whitenoise==3.3.1
